+++ skforecast/plot/__init__.py
@@ -1,1 +1,1 @@
-from .plot import plot_residuals
+from .plot import plot_residuals, plot_multivariate_time_series_corr